+++ tests/python/rest_api/test_jobs.py
@@ -4,22 +4,26 @@
 # SPDX-License-Identifier: MIT
 
 import json
+import os
 import xml.etree.ElementTree as ET
 import zipfile
 from copy import deepcopy
 from http import HTTPStatus
 from io import BytesIO
-from typing import List
+from typing import Any, Dict, List, Optional
 
+import numpy as np
 import pytest
+from cvat_sdk import models
 from cvat_sdk.api_client.api_client import ApiClient, Endpoint
 from cvat_sdk.core.helpers import get_paginated_collection
 from deepdiff import DeepDiff
 from PIL import Image
 
 from shared.utils.config import make_api_client
+from shared.utils.helpers import generate_image_files
 
-from .utils import CollectionSimpleFilterTestBase, export_dataset
+from .utils import CollectionSimpleFilterTestBase, compare_annotations, create_task, export_dataset
 
 
 def get_job_staff(job, tasks, projects):
@@ -52,21 +56,353 @@ def filter_jobs(jobs, tasks, org):
     return jobs, kwargs
 
 
+@pytest.mark.usefixtures("restore_db_per_function")
+class TestPostJobs:
+    def _test_create_job_ok(self, user: str, data: Dict[str, Any], **kwargs):
+        with make_api_client(user) as api_client:
+            (_, response) = api_client.jobs_api.create(
+                models.JobWriteRequest(**deepcopy(data)), **kwargs
+            )
+            assert response.status == HTTPStatus.CREATED
+        return response
+
+    def _test_create_job_fails(
+        self, user: str, data: Dict[str, Any], *, expected_status: int, **kwargs
+    ):
+        with make_api_client(user) as api_client:
+            (_, response) = api_client.jobs_api.create(
+                models.JobWriteRequest(**deepcopy(data)),
+                **kwargs,
+                _check_status=False,
+                _parse_response=False,
+            )
+            assert response.status == expected_status
+        return response
+
+    @pytest.mark.parametrize("task_mode", ["annotation", "interpolation"])
+    def test_can_create_gt_job_with_manual_frames(self, admin_user, tasks, task_mode):
+        user = admin_user
+        job_frame_count = 4
+        task = next(
+            t
+            for t in tasks
+            if not t["project_id"]
+            and not t["organization"]
+            and t["mode"] == task_mode
+            and t["size"] > job_frame_count
+        )
+        task_id = task["id"]
+        with make_api_client(user) as api_client:
+            (task_meta, _) = api_client.tasks_api.retrieve_data_meta(task_id)
+            frame_step = int(task_meta.frame_filter.split("=")[-1]) if task_meta.frame_filter else 1
+
+        job_frame_ids = list(range(task_meta.start_frame, task_meta.stop_frame, frame_step))[
+            :job_frame_count
+        ]
+        job_spec = {
+            "task_id": task_id,
+            "type": "ground_truth",
+            "frame_selection_method": "manual",
+            "frames": job_frame_ids,
+        }
+
+        response = self._test_create_job_ok(user, job_spec)
+        job_id = json.loads(response.data)["id"]
+
+        with make_api_client(user) as api_client:
+            (gt_job_meta, _) = api_client.jobs_api.retrieve_data_meta(job_id)
+
+        assert job_frame_count == gt_job_meta.size
+        assert job_frame_ids == gt_job_meta.included_frames
+
+    @pytest.mark.parametrize("task_mode", ["annotation", "interpolation"])
+    def test_can_create_gt_job_with_random_frames(self, admin_user, tasks, task_mode):
+        user = admin_user
+        job_frame_count = 3
+        required_task_frame_count = job_frame_count + 1
+        task = next(
+            t
+            for t in tasks
+            if not t["project_id"]
+            and not t["organization"]
+            and t["mode"] == task_mode
+            and t["size"] > required_task_frame_count
+        )
+        task_id = task["id"]
+
+        job_spec = {
+            "task_id": task_id,
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": job_frame_count,
+        }
+
+        response = self._test_create_job_ok(user, job_spec)
+        job_id = json.loads(response.data)["id"]
+
+        with make_api_client(user) as api_client:
+            (gt_job_meta, _) = api_client.jobs_api.retrieve_data_meta(job_id)
+
+        assert job_frame_count == gt_job_meta.size
+        assert job_frame_count == len(gt_job_meta.included_frames)
+
+    @pytest.mark.parametrize(
+        "task_id, frame_ids",
+        [
+            # The results have to be the same in different CVAT revisions,
+            # so the task ids are fixed
+            (21, [3, 5, 7]),  # annotation task
+            (5, [11, 14, 20]),  # interpolation task
+        ],
+    )
+    def test_can_create_gt_job_with_random_frames_and_seed(self, admin_user, task_id, frame_ids):
+        user = admin_user
+        job_spec = {
+            "task_id": task_id,
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 3,
+            "seed": 42,
+        }
+
+        response = self._test_create_job_ok(user, job_spec)
+        job_id = json.loads(response.data)["id"]
+
+        with make_api_client(user) as api_client:
+            (gt_job_meta, _) = api_client.jobs_api.retrieve_data_meta(job_id)
+
+        assert gt_job_meta.included_frames == frame_ids
+
+    def test_can_create_no_more_than_1_gt_job(self, admin_user, jobs):
+        user = admin_user
+        task_id = next(j for j in jobs if j["type"] == "ground_truth")["task_id"]
+
+        job_spec = {
+            "task_id": task_id,
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        response = self._test_create_job_fails(
+            user, job_spec, expected_status=HTTPStatus.BAD_REQUEST
+        )
+        assert b"A task can have only 1 ground truth job" in response.data
+
+    def test_can_create_gt_job_in_sandbox_task(self, tasks, jobs, users):
+        task = next(
+            t
+            for t in tasks
+            if t["organization"] is None
+            and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+            and not users[t["owner"]["id"]]["is_superuser"]
+        )
+        user = task["owner"]["username"]
+
+        job_spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        self._test_create_job_ok(user, job_spec)
+
+    @pytest.mark.parametrize(
+        "org_role, is_staff, allow",
+        [
+            ("owner", True, True),
+            ("owner", False, True),
+            ("maintainer", True, True),
+            ("maintainer", False, True),
+            ("supervisor", True, True),
+            ("supervisor", False, False),
+            ("worker", True, False),
+            ("worker", False, False),
+        ],
+    )
+    def test_create_gt_job_in_org_task(
+        self, tasks, jobs, users, is_org_member, is_task_staff, org_role, is_staff, allow
+    ):
+        for user in users:
+            if user["is_superuser"]:
+                continue
+
+            task = next(
+                (
+                    t
+                    for t in tasks
+                    if t["organization"] is not None
+                    and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+                    and is_task_staff(user["id"], t["id"]) == is_staff
+                    and is_org_member(user["id"], t["organization"], role=org_role)
+                ),
+                None,
+            )
+            if task is not None:
+                break
+
+        assert task
+
+        job_spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        if allow:
+            self._test_create_job_ok(user["username"], job_spec)
+        else:
+            self._test_create_job_fails(
+                user["username"], job_spec, expected_status=HTTPStatus.FORBIDDEN
+            )
+
+    def test_create_response_matches_get(self, tasks, jobs, users):
+        task = next(
+            t
+            for t in tasks
+            if t["organization"] is None
+            and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+            and not users[t["owner"]["id"]]["is_superuser"]
+        )
+        user = task["owner"]["username"]
+
+        spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        response = self._test_create_job_ok(user, spec)
+        job = json.loads(response.data)
+
+        with make_api_client(user) as api_client:
+            (_, response) = api_client.jobs_api.retrieve(job["id"])
+            assert DeepDiff(job, json.loads(response.data), ignore_order=True) == {}
+
+
+@pytest.mark.usefixtures("restore_db_per_function")
+class TestDeleteJobs:
+    def _test_destroy_job_ok(self, user, job_id, **kwargs):
+        with make_api_client(user) as api_client:
+            (_, response) = api_client.jobs_api.destroy(job_id, **kwargs)
+            assert response.status == HTTPStatus.NO_CONTENT
+
+    def _test_destroy_job_fails(self, user, job_id, *, expected_status: int, **kwargs):
+        with make_api_client(user) as api_client:
+            (_, response) = api_client.jobs_api.destroy(
+                job_id, **kwargs, _check_status=False, _parse_response=False
+            )
+            assert response.status == expected_status
+        return response
+
+    @pytest.mark.usefixtures("restore_cvat_data")
+    @pytest.mark.parametrize("job_type, allow", (("ground_truth", True), ("annotation", False)))
+    def test_destroy_job(self, admin_user, jobs, job_type, allow):
+        job = next(j for j in jobs if j["type"] == job_type)
+
+        if allow:
+            self._test_destroy_job_ok(admin_user, job["id"])
+        else:
+            self._test_destroy_job_fails(
+                admin_user, job["id"], expected_status=HTTPStatus.BAD_REQUEST
+            )
+
+    def test_can_destroy_gt_job_in_sandbox_task(self, tasks, jobs, users, admin_user):
+        task = next(
+            t
+            for t in tasks
+            if t["organization"] is None
+            and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+            and not users[t["owner"]["id"]]["is_superuser"]
+        )
+        user = task["owner"]["username"]
+
+        job_spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        with make_api_client(admin_user) as api_client:
+            (job, _) = api_client.jobs_api.create(job_spec)
+
+        self._test_destroy_job_ok(user, job.id)
+
+    @pytest.mark.parametrize(
+        "org_role, is_staff, allow",
+        [
+            ("owner", True, True),
+            ("owner", False, True),
+            ("maintainer", True, True),
+            ("maintainer", False, True),
+            ("supervisor", True, True),
+            ("supervisor", False, False),
+            ("worker", True, False),
+            ("worker", False, False),
+        ],
+    )
+    def test_destroy_gt_job_in_org_task(
+        self,
+        tasks,
+        jobs,
+        users,
+        is_org_member,
+        is_task_staff,
+        org_role,
+        is_staff,
+        allow,
+        admin_user,
+    ):
+        for user in users:
+            task = next(
+                (
+                    t
+                    for t in tasks
+                    if t["organization"] is not None
+                    and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+                    and is_task_staff(user["id"], t["id"]) == is_staff
+                    and is_org_member(user["id"], t["organization"], role=org_role)
+                ),
+                None,
+            )
+            if task is not None:
+                break
+
+        assert task
+
+        job_spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        with make_api_client(admin_user) as api_client:
+            (job, _) = api_client.jobs_api.create(job_spec)
+
+        if allow:
+            self._test_destroy_job_ok(user["username"], job.id)
+        else:
+            self._test_destroy_job_fails(
+                user["username"], job.id, expected_status=HTTPStatus.FORBIDDEN
+            )
+
+
 @pytest.mark.usefixtures("restore_db_per_class")
 class TestGetJobs:
-    def _test_get_job_200(self, user, jid, data, **kwargs):
+    def _test_get_job_200(
+        self, user, jid, *, expected_data: Optional[Dict[str, Any]] = None, **kwargs
+    ):
         with make_api_client(user) as client:
             (_, response) = client.jobs_api.retrieve(jid, **kwargs)
             assert response.status == HTTPStatus.OK
-            assert (
-                DeepDiff(
-                    data,
-                    json.loads(response.data),
-                    exclude_paths="root['updated_date']",
-                    ignore_order=True,
-                )
-                == {}
-            )
+
+            if expected_data is not None:
+                assert compare_annotations(expected_data, json.loads(response.data)) == {}
 
     def _test_get_job_403(self, user, jid, **kwargs):
         with make_api_client(user) as client:
@@ -75,13 +411,13 @@ class TestGetJobs:
             )
             assert response.status == HTTPStatus.FORBIDDEN
 
-    def test_admin_can_get_sandbox_job(self, jobs, tasks):
+    def test_admin_can_get_sandbox_job(self, admin_user, jobs, tasks):
         job = next(job for job in jobs if tasks[job["task_id"]]["organization"] is None)
-        self._test_get_job_200("admin2", job["id"], job)
+        self._test_get_job_200(admin_user, job["id"], expected_data=job)
 
-    def test_admin_can_get_org_job(self, jobs, tasks):
+    def test_admin_can_get_org_job(self, admin_user, jobs, tasks):
         job = next(job for job in jobs if tasks[job["task_id"]]["organization"] is not None)
-        self._test_get_job_200("admin2", job["id"], job)
+        self._test_get_job_200(admin_user, job["id"], expected_data=job)
 
     @pytest.mark.parametrize("groups", [["business"], ["user"]])
     def test_non_admin_org_staff_can_get_job(
@@ -94,7 +430,7 @@ class TestGetJobs:
             if user["groups"] == groups and user["id"] in org_staff(org["id"])
         )
         job = jobs_by_org[org_id][0]
-        self._test_get_job_200(user["username"], job["id"], job)
+        self._test_get_job_200(user["username"], job["id"], expected_data=job)
 
     @pytest.mark.parametrize("groups", [["business"], ["user"], ["worker"]])
     def test_non_admin_job_staff_can_get_job(self, groups, users, jobs, is_job_staff):
@@ -104,7 +440,7 @@ class TestGetJobs:
             for job in jobs
             if user["groups"] == groups and is_job_staff(user["id"], job["id"])
         )
-        self._test_get_job_200(user["username"], job["id"], job)
+        self._test_get_job_200(user["username"], job["id"], expected_data=job)
 
     @pytest.mark.parametrize("groups", [["business"], ["user"], ["worker"]])
     def test_non_admin_non_job_staff_non_org_staff_cannot_get_job(
@@ -121,6 +457,309 @@ class TestGetJobs:
         )
         self._test_get_job_403(user["username"], job_id)
 
+    @pytest.mark.usefixtures("restore_db_per_function")
+    def test_can_get_gt_job_in_sandbox_task(self, tasks, jobs, users, admin_user):
+        task = next(
+            t
+            for t in tasks
+            if t["organization"] is None
+            and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+            and not users[t["owner"]["id"]]["is_superuser"]
+        )
+        user = task["owner"]["username"]
+
+        job_spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        with make_api_client(admin_user) as api_client:
+            (job, _) = api_client.jobs_api.create(job_spec)
+
+        self._test_get_job_200(user, job.id)
+
+    @pytest.mark.usefixtures("restore_db_per_function")
+    @pytest.mark.parametrize(
+        "org_role, is_staff, allow",
+        [
+            ("owner", True, True),
+            ("owner", False, True),
+            ("maintainer", True, True),
+            ("maintainer", False, True),
+            ("supervisor", True, True),
+            ("supervisor", False, False),
+            ("worker", True, True),
+            ("worker", False, False),
+        ],
+    )
+    def test_get_gt_job_in_org_task(
+        self,
+        tasks,
+        jobs,
+        users,
+        is_org_member,
+        is_task_staff,
+        org_role,
+        is_staff,
+        allow,
+        admin_user,
+    ):
+        for user in users:
+            task = next(
+                (
+                    t
+                    for t in tasks
+                    if t["organization"] is not None
+                    and all(j["type"] != "ground_truth" for j in jobs if j["task_id"] == t["id"])
+                    and is_task_staff(user["id"], t["id"]) == is_staff
+                    and is_org_member(user["id"], t["organization"], role=org_role)
+                ),
+                None,
+            )
+            if task is not None:
+                break
+
+        assert task
+
+        job_spec = {
+            "task_id": task["id"],
+            "type": "ground_truth",
+            "frame_selection_method": "random_uniform",
+            "frame_count": 1,
+        }
+
+        with make_api_client(admin_user) as api_client:
+            (_, response) = api_client.jobs_api.create(job_spec)
+            job = json.loads(response.data)
+
+        if allow:
+            self._test_get_job_200(user["username"], job["id"], expected_data=job)
+        else:
+            self._test_get_job_403(user["username"], job["id"])
+
+
+@pytest.mark.usefixtures(
+    # if the db is restored per test, there are conflicts with the server data cache
+    # if we don't clean the db, the gt jobs created will be reused, and their
+    # ids won't conflict
+    "restore_db_per_class"
+)
+class TestGetGtJobData:
+    @pytest.mark.usefixtures("restore_db_per_function")
+    @pytest.mark.parametrize("task_mode", ["annotation", "interpolation"])
+    def test_can_get_gt_job_meta(self, admin_user, tasks, task_mode):
+        user = admin_user
+        job_frame_count = 4
+        task = next(
+            t
+            for t in tasks
+            if not t["project_id"]
+            and not t["organization"]
+            and t["mode"] == task_mode
+            and t["size"] > job_frame_count
+        )
+        task_id = task["id"]
+        with make_api_client(user) as api_client:
+            (task_meta, _) = api_client.tasks_api.retrieve_data_meta(task_id)
+            frame_step = int(task_meta.frame_filter.split("=")[-1]) if task_meta.frame_filter else 1
+
+        job_frame_ids = list(range(task_meta.start_frame, task_meta.stop_frame, frame_step))[
+            :job_frame_count
+        ]
+        gt_job = self._get_or_create_gt_job(admin_user, task_id, job_frame_ids)
+
+        with make_api_client(user) as api_client:
+            (gt_job_meta, _) = api_client.jobs_api.retrieve_data_meta(gt_job.id)
+
+        # These values are relative to the resulting task frames, unlike meta values
+        assert 0 == gt_job.start_frame
+        assert task_meta.size - 1 == gt_job.stop_frame
+
+        # The size is adjusted by the frame step and included frames
+        assert job_frame_count == gt_job_meta.size
+        assert job_frame_ids == gt_job_meta.included_frames
+
+        # The frames themselves are the same as in the whole range
+        # this is required by the UI implementation
+        assert task_meta.start_frame == gt_job_meta.start_frame
+        assert task_meta.stop_frame == gt_job_meta.stop_frame
+        if task_mode == "annotation":
+            assert (
+                len(gt_job_meta.frames)
+                == (gt_job_meta.stop_frame + 1 - gt_job_meta.start_frame) / frame_step
+            )
+        elif task_mode == "interpolation":
+            assert len(gt_job_meta.frames) == 1
+        else:
+            assert False
+
+    @pytest.mark.usefixtures("restore_db_per_function")
+    def test_can_get_gt_job_meta_with_complex_frame_setup(self, admin_user):
+        image_count = 50
+        start_frame = 3
+        stop_frame = image_count - 4
+        frame_step = 5
+
+        images = generate_image_files(image_count)
+
+        task_id, _ = create_task(
+            admin_user,
+            spec={
+                "name": "test complex frame setup",
+                "labels": [{"name": "cat"}],
+            },
+            data={
+                "image_quality": 75,
+                "start_frame": start_frame,
+                "stop_frame": stop_frame,
+                "frame_filter": f"step={frame_step}",
+                "client_files": images,
+                "sorting_method": "predefined",
+            },
+        )
+
+        task_frame_ids = range(start_frame, stop_frame, frame_step)
+        job_frame_ids = list(task_frame_ids[::3])
+        gt_job = self._get_or_create_gt_job(admin_user, task_id, job_frame_ids)
+
+        with make_api_client(admin_user) as api_client:
+            (gt_job_meta, _) = api_client.jobs_api.retrieve_data_meta(gt_job.id)
+
+        # These values are relative to the resulting task frames, unlike meta values
+        assert 0 == gt_job.start_frame
+        assert len(task_frame_ids) - 1 == gt_job.stop_frame
+
+        # The size is adjusted by the frame step and included frames
+        assert len(job_frame_ids) == gt_job_meta.size
+        assert job_frame_ids == gt_job_meta.included_frames
+
+        # The frames themselves are the same as in the whole range
+        # with placeholders in the frames outside the job.
+        # This is required by the UI implementation
+        assert start_frame == gt_job_meta.start_frame
+        assert max(task_frame_ids) == gt_job_meta.stop_frame
+        assert [frame_info["name"] for frame_info in gt_job_meta.frames] == [
+            images[frame].name if frame in job_frame_ids else "placeholder.jpg"
+            for frame in task_frame_ids
+        ]
+
+    @pytest.mark.parametrize("task_mode", ["annotation", "interpolation"])
+    @pytest.mark.parametrize("quality", ["compressed", "original"])
+    def test_can_get_gt_job_chunk(self, admin_user, tasks, task_mode, quality):
+        user = admin_user
+        job_frame_count = 4
+        task = next(
+            t
+            for t in tasks
+            if not t["project_id"]
+            and not t["organization"]
+            and t["mode"] == task_mode
+            and t["size"] > job_frame_count
+        )
+        task_id = task["id"]
+        with make_api_client(user) as api_client:
+            (task_meta, _) = api_client.tasks_api.retrieve_data_meta(task_id)
+            frame_step = int(task_meta.frame_filter.split("=")[-1]) if task_meta.frame_filter else 1
+
+        job_frame_ids = list(range(task_meta.start_frame, task_meta.stop_frame, frame_step))[
+            :job_frame_count
+        ]
+        gt_job = self._get_or_create_gt_job(admin_user, task_id, job_frame_ids)
+
+        with make_api_client(admin_user) as api_client:
+            (chunk_file, response) = api_client.jobs_api.retrieve_data(
+                gt_job.id, number=0, quality=quality, type="chunk"
+            )
+            assert response.status == HTTPStatus.OK
+
+        frame_range = range(
+            task_meta.start_frame, min(task_meta.stop_frame + 1, task_meta.chunk_size), frame_step
+        )
+        included_frames = job_frame_ids
+
+        # The frame count is the same as in the whole range
+        # with placeholders in the frames outside the job.
+        # This is required by the UI implementation
+        with zipfile.ZipFile(chunk_file) as chunk:
+            assert set(chunk.namelist()) == set("{:06d}.jpeg".format(i) for i in frame_range)
+
+            for file_info in chunk.filelist:
+                with chunk.open(file_info) as image_file:
+                    image = Image.open(image_file)
+                    image_data = np.array(image)
+
+                if int(os.path.splitext(file_info.filename)[0]) not in included_frames:
+                    assert image.size == (1, 1)
+                    assert np.all(image_data == 0), image_data
+                else:
+                    assert image.size > (1, 1)
+                    assert np.any(image_data != 0)
+
+    def _get_or_create_gt_job(self, user, task_id, frames):
+        with make_api_client(user) as api_client:
+            (task_jobs, _) = api_client.jobs_api.list(task_id=task_id, type="ground_truth")
+            if task_jobs.results:
+                gt_job = task_jobs.results[0]
+            else:
+                job_spec = {
+                    "task_id": task_id,
+                    "type": "ground_truth",
+                    "frame_selection_method": "manual",
+                    "frames": frames,
+                }
+
+                (gt_job, _) = api_client.jobs_api.create(job_spec)
+
+        return gt_job
+
+    @pytest.mark.parametrize("task_mode", ["annotation", "interpolation"])
+    @pytest.mark.parametrize("quality", ["compressed", "original"])
+    def test_can_get_gt_job_frame(self, admin_user, tasks, task_mode, quality):
+        user = admin_user
+        job_frame_count = 4
+        task = next(
+            t
+            for t in tasks
+            if not t["project_id"]
+            and not t["organization"]
+            and t["mode"] == task_mode
+            and t["size"] > job_frame_count
+        )
+        task_id = task["id"]
+        with make_api_client(user) as api_client:
+            (task_meta, _) = api_client.tasks_api.retrieve_data_meta(task_id)
+            frame_step = int(task_meta.frame_filter.split("=")[-1]) if task_meta.frame_filter else 1
+
+        job_frame_ids = list(range(task_meta.start_frame, task_meta.stop_frame, frame_step))[
+            :job_frame_count
+        ]
+        gt_job = self._get_or_create_gt_job(admin_user, task_id, job_frame_ids)
+
+        frame_range = range(
+            task_meta.start_frame, min(task_meta.stop_frame + 1, task_meta.chunk_size), frame_step
+        )
+        included_frames = job_frame_ids
+        excluded_frames = list(set(frame_range).difference(included_frames))
+
+        with make_api_client(admin_user) as api_client:
+            (_, response) = api_client.jobs_api.retrieve_data(
+                gt_job.id,
+                number=excluded_frames[0],
+                quality=quality,
+                type="frame",
+                _parse_response=False,
+                _check_status=False,
+            )
+            assert response.status == HTTPStatus.BAD_REQUEST
+            assert b"The frame number doesn't belong to the job" in response.data
+
+            (_, response) = api_client.jobs_api.retrieve_data(
+                gt_job.id, number=included_frames[0], quality=quality, type="frame"
+            )
+            assert response.status == HTTPStatus.OK
+
 
 @pytest.mark.usefixtures("restore_db_per_class")
 class TestListJobs:
@@ -129,10 +768,7 @@ class TestListJobs:
             results = get_paginated_collection(
                 client.jobs_api.list_endpoint, return_json=True, **kwargs
             )
-            assert (
-                DeepDiff(data, results, exclude_paths="root['updated_date']", ignore_order=True)
-                == {}
-            )
+            assert compare_annotations(data, results) == {}
 
     def _test_list_jobs_403(self, user, **kwargs):
         with make_api_client(user) as client:
@@ -188,6 +824,7 @@ class TestJobsListFilters(CollectionSimpleFilterTestBase):
             "stage",
             "task_id",
             "project_id",
+            "type",
         ),
     )
     def test_can_use_simple_filter_for_object_list(self, field):
@@ -200,12 +837,7 @@ class TestGetAnnotations:
         with make_api_client(user) as client:
             (_, response) = client.jobs_api.retrieve_annotations(jid)
             assert response.status == HTTPStatus.OK
-
-            response_data = json.loads(response.data)
-            assert (
-                DeepDiff(data, response_data, exclude_regex_paths=r"root\['version|updated_date'\]")
-                == {}
-            )
+            assert compare_annotations(data, json.loads(response.data)) == {}
 
     def _test_get_job_annotations_403(self, user, jid):
         with make_api_client(user) as client:
@@ -311,6 +943,13 @@ class TestGetAnnotations:
         else:
             self._test_get_job_annotations_403(username, job_id)
 
+    @pytest.mark.parametrize("job_type", ("ground_truth", "annotation"))
+    def test_can_get_annotations(self, admin_user, jobs, annotations, job_type):
+        job = next(j for j in jobs if j["type"] == job_type)
+        self._test_get_job_annotations_200(
+            admin_user, job["id"], annotations["job"][str(job["id"])]
+        )
+
 
 @pytest.mark.usefixtures("restore_db_per_function")
 class TestPatchJobAnnotations:
@@ -326,14 +965,7 @@ class TestPatchJobAnnotations:
 
             if expect_success:
                 assert response.status == HTTPStatus.OK
-                assert (
-                    DeepDiff(
-                        data,
-                        json.loads(response.data),
-                        exclude_regex_paths=r"root\['version|updated_date'\]",
-                    )
-                    == {}
-                )
+                assert compare_annotations(data, json.loads(response.data)) == {}
             else:
                 assert response.status == HTTPStatus.FORBIDDEN
 
@@ -441,6 +1073,12 @@ class TestPatchJobAnnotations:
 
         data = request_data(jid)
         self._check_response(username, jid, expect_success, data)
+
+    @pytest.mark.parametrize("job_type", ("ground_truth", "annotation"))
+    def test_can_update_annotations(self, admin_user, jobs, request_data, job_type):
+        job = next(j for j in jobs if j["type"] == job_type)
+        data = request_data(job["id"])
+        self._check_response(admin_user, job["id"], True, data)
 
 
 @pytest.mark.usefixtures("restore_db_per_function")
@@ -813,15 +1451,15 @@ class TestGetJobPreview:
 
 @pytest.mark.usefixtures("restore_db_per_class")
 class TestGetJobDataMeta:
-    def test_admin_can_get_job_meta(self, jobs):
-        with make_api_client("admin1") as client:
-            job_id = jobs.raw[0]["id"]
+    @pytest.mark.parametrize("org_slug", [None, "", "org"])
+    def test_can_get_job_meta_with_org_slug(self, admin_user, tasks, jobs, organizations, org_slug):
+        # Checks for backward compatibility with org_slug parameter
+        task = next(t for t in tasks if t["organization"])
+        job = next(j for j in jobs if j["task_id"] == task["id"])
 
-            client.organization_slug = None
-            client.jobs_api.retrieve_data_meta(job_id)
+        if org_slug == "org":
+            org_slug = organizations[task["organization"]]["slug"]
 
-            client.organization_slug = ""
-            client.jobs_api.retrieve_data_meta(job_id)
-
-            client.organization_slug = "org1"
-            client.jobs_api.retrieve_data_meta(job_id)
+        with make_api_client(admin_user) as client:
+            client.organization_slug = org_slug
+            client.jobs_api.retrieve_data_meta(job["id"])